refactor: 예약 상태 전체 목록 조회 기능를 관리자 전용 기능으로 수정
