+++ src/main/resources/static/js/reservation-with-member.js
@@ -3,7 +3,7 @@ const RESERVATION_API_ENDPOINT = '/admin/reservations';
 const TIME_API_ENDPOINT = '/times';
 const THEME_API_ENDPOINT = '/themes';
 const MEMBER_API_ENDPOINT = '/members';
-const STATUS_API_ENDPOINT = '/reservations/statuses';
+const STATUS_API_ENDPOINT = '/admin/reservations/statuses';
 const timesOptions = [];
 const themesOptions = [];
 const membersOptions = [];
